pyproject.toml
[project]
name = "zatom"
version = "0.0.1"
description = "Official repository of Zatom, a multimodal all-atom generative model"
authors = [
    { name = "Alex Morehead", email = "alex.morehead@gmail.edu" }
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = "MIT"
keywords = [
    'artificial intelligence',
    'deep learning',
    '3D molecules',
]

classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.10',
]

dependencies = [
    # --------- pytorch --------- #
    "torch==2.8.0",
    "torchvision==0.23",
    "lightning==2.5.3",
    "torchmetrics==1.8.1",
    "lovely_tensors==0.1.18",
    "einops==0.8.1",
    "einx==0.3.0",
    "jaxtyping==0.3.2",
    "torch_geometric==2.6.1",
    "jvp_flash_attention==0.10.0",

    # --------- torch-scatter --------- #
    "torch-scatter @ https://data.pyg.org/whl/torch-2.8.0%2Bcpu/torch_scatter-2.1.2%2Bpt28cpu-cp310-cp310-linux_x86_64.whl ; sys_platform == 'linux'",
    "torch-scatter @ https://data.pyg.org/whl/torch-2.8.0%2Bcpu/torch_scatter-2.1.2%2Bpt28cpu-cp310-cp310-win_amd64.whl ; sys_platform == 'win32'",
    "torch-scatter @ https://data.pyg.org/whl/torch-2.8.0%2Bcpu/torch_scatter-2.1.2-cp310-cp310-macosx_10_9_universal2.whl ; sys_platform == 'darwin'",

    # --------- configs --------- #
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",

    # --------- others --------- #
    "beartype==0.21.0",
    "environs==14.3.0",
    "pre-commit==4.3.0",
    "numpy==2.2.6",
    "rdkit==2025.3.5",
    "dm-tree==0.1.8",
    "requests==2.32.3",
    "rich==14.1.0",
    "rootutils==1.0.7",
    "pandas==2.3.1",
    "pytest==8.4.1",
    "types-requests==2.32.4.20250809",
    "fairscale==0.4.13",
    "mashumaro==3.14",
    "modelcif==1.2",
    "networkx==3.3",
    "wandb==0.18.7",
    "click==8.1.7",
    "pyyaml==6.0.2",
    "biopython==1.84",
    "scipy==1.13.1",
    "sympy==1.13.3",
    "numba==0.61.2",
    "gemmi==0.6.5",
    "scikit-learn==1.6.1",
    "chembl_structure_pipeline==1.2.2",
    "CairoSVG==2.7.1",
    "ase==3.26.0",
    "e3nn==0.5.1",
    "lmdb==1.7.3",
    "matminer==0.9.3",
    "mofchecker==0.9.6",
    "openbabel-wheel==3.1.1.22",
    "p_tqdm==1.4.1",
    "pathos==0.3.4",
    "posebusters==0.5.1",
    "pymatgen==2024.11.13",
    "pythreejs==2.4.2",
    "pyxtal==0.6.7",
    "reportlab==4.4.3",
    "SMACT==2.8",
    "svglib==1.5.1",
    "svgutils==0.3.4",
    "svgwrite==1.4.3",
    "huggingface_hub==0.34.4",
    "importlib_resources==6.5.2",
    "setuptools==80.9.0",
    "chgnet==0.4.2",
    "geoopt==0.5.1",
    "chemparse==0.3.2",
    "toolz==1.1.0",
    "geopandas==1.1.1",
    "seaborn==0.13.2",
    "libigl==2.6.1",
    "flow_matching @ git+https://github.com/amorehead/flow_matching.git@mm",
    "fairchem-core @ git+https://github.com/amorehead/fairchem.git@torch-2.6.0-plus#subdirectory=packages/fairchem-core",
    "platonic_transformers @ git+https://github.com/amorehead/PlatonicTransformers.git@packaging",
]

[project.urls]
Homepage = "https://pypi.org/project/zatom/"
Repository = "https://github.com/amorehead/zatom"

[project.optional-dependencies]
test = ["pytest", "requests"]
cuda = [
    "cuequivariance_ops_cu12==0.6.0",
    "cuequivariance_ops_torch_cu12==0.6.0",
    "cuequivariance_torch==0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

# Assuming you're developing for Python 3.10
target-version = "py310"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["zatom"]

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
